update plot, support bfloat16

+++ vllm_omni/diffusion/attention/layer.py
@@ -98,7 +98,7 @@ class Attention(nn.Module):
                 os.makedirs(save_dir, exist_ok=True)
                 
                 plt.figure(figsize=(10, 10))
-                plt.imshow(attn_weights[0, 0].detach().cpu().numpy(), cmap='viridis')
+                plt.imshow(attn_weights[0, 0].detach().float().cpu().numpy(), cmap='viridis')
                 plt.colorbar()
                 plt.title(f"Attention Map Step {step}")
                 plt.savefig(os.path.join(save_dir, f"attn_map_step_{step:03d}.png"))
